@@ -1,59 +1,42 @@
-import { Component } from 'react';
-import Spinner from '../spinner/spinner'
-import ErrorMessage from '../errorMessage/errorMessage';
-import MarvelService from '../../services/MarvelService';
+import { useState, useEffect } from 'react';
+import Spinner from '../spinner/Spinner'
+import ErrorMessage from '../errorMessage/ErrorMessage';
+import useMarvelService from '../../services/MarvelService';
 import './randomChar.scss';
 import mjolnir from '../../resources/img/mjolnir.png';
 
-class RandomChar extends Component {
-    state = {
-        char: {},
-        loading: true,
-        error: false
+const RandomChar = () => {
+
+    const [char, setChar] = useState(null);
+
+    const {loading, error, getCharacter, clearError} = useMarvelService();
+
+    useEffect(() => {
+        updateChar();
+        // timerId = setInterval(this.updateChar, 3000)
+        // return () => {
+        //     clearInterval(timerId)
+        // }
+    }, [])
+
+
+    const onCharLoaded = (char) => {
+        setChar(char);
     }
 
-    marvelService = new MarvelService();
 
-    componentDidMount() {
-        this.updateChar();
-        // this.timerId = setInterval(this.updateChar, 3000)
-    }
 
-    componentWillUnmount() {
-        clearInterval(this.timerId)
-    }
-
-    onCharLoaded = (char) => {
-        this.setState({
-            char,
-            loading: false,
-        })
-    }
-
-    onError = (char) => {
-        this.setState({
-            loading: false,
-            error: true})
-    }
-
-    updateChar = () => {
+    const updateChar = () => {
+        clearError();
         const id = Math.floor(Math.random() * (1011400 - 1011000) + 1011000);
 
-        this.setState({
-            loading: true
-        })
- 
-        this.marvelService
-        .getCharacter(id)
-            .then(this.onCharLoaded)
-            .catch(this.onError);
-    }
+        getCharacter(id)
+            .then(onCharLoaded);
+        }
 
-    render() {
-        const {char, loading, error} = this.state;        
         const errorMessage = error ? <ErrorMessage/> : null;
         const spinner = loading ? <Spinner/> : null;
-        const content = !(loading || error) ? <View char={char}/> : null
+        const content = !(loading || error || !char) ? <View char={char}/> : null
 
         return (
             <div className="randomchar">
@@ -69,14 +52,13 @@ class RandomChar extends Component {
                         Or choose another one
                     </p>
                     <button className="button button__main"
-                            onClick={this.updateChar}>
+                            onClick={updateChar}>
                         <div className="inner">try it</div>
                     </button>
                     <img src={mjolnir} alt="mjolnir" className="randomchar__decoration"/>
                 </div>
             </div>
         )
-    }
 }
 
 const View = ({char}) => {
