@@ -1,48 +1,47 @@
-import { Component } from "react/cjs/react.production.min";
+import {lazy, Suspense} from 'react';
+import { BrowserRouter as Router, Route, Routes} from "react-router-dom";
 
 import AppHeader from "../appHeader/AppHeader";
-import RandomChar from "../randomChar/RandomChar";
-import CharList from "../charList/CharList";
-import CharInfo from "../charInfo/CharInfo";
+import Spinner from '../spinner/Spinner';
 
-import decoration from '../../resources/img/vision.png';
-import ErrorBoundary from "../errorBoundary/ErrorBoundary";
+const Page404 = lazy(() => import('../pages/404'));
+const MainPage = lazy(() => import('../pages/MainPage'));
+const ComicsPage = lazy(() => import('../pages/ComicsPage'));
+const SingleComicLayout = lazy(() => import('../pages/singleComicLayout/SingleComicLayout'));
+const SingleCharacterLayout = lazy(() => import('../pages/singleCharacterLayout/SingleCharacterLayout'));
+const SinglePage = lazy(() => import('../pages/SinglePage'));
 
-class App extends Component{
 
-    state = {
-        selectedChar: null
-    }
+const App = () => {
 
-    onCharSelected = (id) => {
-        this.setState({
-            selectedChar: id
-        })
-    }
-
-    render() {
-        return (
+    return (
+        <Router>
             <div className="app">
                 <AppHeader/>
                 <main>
-                    <ErrorBoundary>
-                        <RandomChar/>
-                    </ErrorBoundary>
-                    <div className="char__content">
-                        <ErrorBoundary>
-                            <CharList
-                            onCharSelected={this.onCharSelected}/>
-                        </ErrorBoundary>
-                        <ErrorBoundary>
-                            <CharInfo
-                            charId={this.state.selectedChar}/>
-                        </ErrorBoundary>
-                    </div>
-                    <img className="bg-decoration" src={decoration} alt="vision"/>
+                    <Suspense fallback={<Spinner/>}>
+                        <Routes>
+                            <Route 
+                                end path="/" 
+                                element={<MainPage/>}/>
+                            <Route 
+                                end path="/comics" 
+                                element={<ComicsPage/>}/>
+                            <Route 
+                                end path="/comics/:id"
+                                element={<SinglePage Component={SingleComicLayout} dataType='comic'/>}/>
+                            <Route 
+                                end path="/characters/:id"
+                                element={<SinglePage Component={SingleCharacterLayout} dataType='character'/>}/>                            
+                            <Route 
+                                end path="*" 
+                                element={<Page404/>}/>
+                        </Routes>
+                    </Suspense>
                 </main>
             </div>
-        )
-    }
+        </Router>
+    )
 }
 
 export default App;
