@@ -1,41 +1,29 @@
-import React, { Component } from 'react/cjs/react.production.min';
+import React, { useState, useEffect, useRef } from 'react';
 import PropTypes from 'prop-types';
-import MarvelService from '../../services/MarvelService';
-import ErrorMessage from '../errorMessage/errorMessage';
-import Spinner from '../spinner/spinner';
-import './charList.scss';
+import useMarvelService from '../../services/MarvelService';
+import ErrorMessage from '../errorMessage/ErrorMessage';
+import Spinner from '../spinner/Spinner';
+import {CSSTransition, TransitionGroup} from 'react-transition-group';import './charList.scss';
 
+const CharList = (props) => {
+    
+    const [charList, setCharList] = useState([]);
+    const [newItemLoading, setNewItemLoading] = useState(false);
+    const [offset, setOffset] = useState(210);
+    const [charEnded, setCharEnded] = useState(false);
 
+    const {loading, error, getAllCharacters} = useMarvelService();
 
+    useEffect(() => {
+        onRequest(offset, true);
+    }, [])
 
-class CharList extends Component {
-    state = {
-        charList: [],
-        loading: true,
-        error: false,
-        newItemLoading: false,
-        offset: 210,
-        charEnded: false
+    const onRequest = (offset, initial) => {
+        initial ? setNewItemLoading(false) : setNewItemLoading(true);
+        updateList(offset);
     }
 
-    marvelService = new MarvelService();
-
-    componentDidMount() {
-        this.updateList();
-    }
-
-    onRequest = (offset) => {
-        this.onCharListLoading();
-        this.updateList(offset);
-    }
-
-    onCharListLoading = () => {
-        this.setState({
-            newItemLoading: true
-        })
-    }
-
-    onCharListLoaded = (newCharList) => {
+    const onCharListLoaded = (newCharList) => {
 
         let ended = false;
 
@@ -43,44 +31,30 @@ class CharList extends Component {
             ended = true;
         }
 
-        this.setState(({charList, offset}) => (
-            {
-                charList: [...charList, ...newCharList],
-                loading: false,
-                newItemLoading: false,
-                offset: offset + 9,
-                charEnded: ended
-            }
-        ))
+        setCharList(charList => [...charList, ...newCharList]);
+        setNewItemLoading(newItemLoading => false);
+        setOffset(offset => offset + 9);
+        setCharEnded(charEnded => ended)
+
+
     }
 
-    updateList = (offset) => {
-        this.marvelService
-            .getAllCharacters(offset)
-            .then(this.onCharListLoaded)
-            .catch(this.onError)
+    const updateList = (offset) => {
+        getAllCharacters(offset)
+            .then(onCharListLoaded);
     }
 
-    onError = () => {
-        this.setState({
-            loading: false,
-            error: true})
+    const itemRefs = useRef([]);
+
+    const focusOnItem = (id) => {
+
+        itemRefs.current.forEach(item => item.classList.remove('char__item_selected'));
+        itemRefs.current[id].classList.add('char__item_selected');
+        itemRefs.current[id].focus();
     }
 
-    itemRefs = [];
+    function renderItems(arr) {
 
-    setRef = (ref) => {
-        this.itemRefs.push(ref);
-    }
-
-    focusOnItem = (id) => {
-
-        this.itemRefs.forEach(item => item.classList.remove('char__item_selected'));
-        this.itemRefs[id].classList.add('char__item_selected');
-        this.itemRefs[id].focus();
-    }
-
-    renderItems(arr) {
         const items =  arr.map((item, i) => {
             let imgStyle = {'objectFit' : 'cover'};
             if (item.thumbnail === 'http://i.annihil.us/u/prod/marvel/i/mg/b/40/image_not_available.jpg') {
@@ -88,52 +62,53 @@ class CharList extends Component {
             }
             
             return (
-                <li 
-                    className="char__item"
-                    key={item.id}
-                    tabIndex={0}
-                    ref={this.setRef}
-                    onClick={() => {
-                        this.props.onCharSelected(item.id);
-                        this.focusOnItem(i);}}>
-                        <img src={item.thumbnail} alt={item.name} style={imgStyle}/>
-                        <div className="char__name">{item.name}</div>
-                </li>
+                <CSSTransition key={i} classNames="char__item" timeout={500}>
+                    <li 
+                        className="char__item"
+                        tabIndex={0}
+                        ref={el => itemRefs.current[i] = el}
+                        onClick={() => {
+                            props.onCharSelected(item.id);
+                            focusOnItem(i);}}
+                        onKeyPress={(e) => {
+                            if (e.key === ' ' || e.key === "Enter") {
+                                props.onCharSelected(item.id);
+                                focusOnItem(i);
+                            }
+                        }}>
+                            <img src={item.thumbnail} alt={item.name} style={imgStyle}/>
+                            <div className="char__name">{item.name}</div>
+                    </li>                    
+                </CSSTransition>
             )
         });
         // А эта конструкция вынесена для центровки спиннера/ошибки
         return (
             <ul className="char__grid">
-                {items}
+                <TransitionGroup component={null}>
+                    {items}
+                </TransitionGroup>
             </ul>
         )
     }
-
-    render() {
-
-        const {charList, loading, error, newItemLoading, offset, charEnded} = this.state;
-        
-        const items = this.renderItems(charList);
-
+        const items = renderItems(charList);
         const errorMessage = error ? <ErrorMessage/> : null;
-        const spinner = loading ? <Spinner/> : null;
-        const content = !(loading || error) ? items : null;
+        const spinner = loading && !newItemLoading ? <Spinner/> : null;
 
         return (
             <div className="char__list">
                 {errorMessage}
                 {spinner}
-                {content}
+                {items}
                 <button 
                 className="button button__main button__long"
                 disabled={newItemLoading}
                 style={{'display': charEnded ? 'none' : 'block'}}
-                onClick={() => this.onRequest(offset)}>
+                onClick={() => onRequest(offset)}>
                     <div className="inner">load more</div>
                 </button>
             </div>
         )
-    }
 }
 
 CharList.propTypes = {
